requirements.txt
groq
openai>=1.50
anthropic>=0.39
requests
httpx>=0.24
ruff>=0.1.0
mypy>=1.0
python-dotenv
fastapi
uvicorn[standard]
slowapi
numpy<2
psutil
stripe
sentry-sdk[fastapi]
prometheus-client
cryptography>=41.0.0
pyyaml>=6.0
Pillow>=10.0.0
